ci: add a job gradle-no-failures to aggregate all build job results (#347)
The gradle-no-failures CI job will pass if no build job failed (i.e.,
all build jobs succeeded or were skipped due to being a duplicate build
that already succeeded).

This should allow us to block PR merging in a way that is compatible
with our usage of the `fkirc/skip-duplicate-actions` action, which skips
a build if it has succeeded elsewhere in CI. Previously the skipping of
previously successful build jobs interfered with making the build jobs
required to merge. This new job alleviates that problem by letting it be
the blocking action instead of the (possibly skipped) build jobs
themselves.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.